fix(HexagonalTest): move main into infrastructure
